+++ frontend/src/components/therapists/TherapistTableRow.jsx
@@ -18,6 +18,12 @@ const TableRow = ({ therapist, fetchTherapists }) => {
   };
 
   const deleteRow = async () => {
+
+    const confirmed = window.confirm("Are you sure you want to delete this patient?");
+    if (!confirmed) {
+      return; // Exit the function if the user cancels the action
+    }
+    
     try {
       const URL = import.meta.env.VITE_API_URL + "therapists/" + therapist.therapistID;
       const response = await axios.delete(URL);
